auth-service/requirements.txt
Django
psycopg2-binary
redis
djangorestframework
django-cors-headers
python-decouple
django-redis
djangorestframework-simplejwt